Crate chalk

Structs

Args
This struct represents the various command line options available.

LoadedProgram
A loaded and parsed program.

Constants

USAGE

Functions

help
Print out help for commands in interpreter mode.

load_program
Load the file into a string, and parse it.

main

process
Process a single command. args is a struct containing the command-line arguments, and prog is a parsed .chalk file.

read_program
Read a program from the command-line. Stop reading when EOF is read. If an error occurs while reading, a Err is returned.

readline_loop
Reads input lines from the user. Lines start with the string given by prompt. Each line the user enters is passed to the function f for processing.

run

Type Definitions

Result