Function chalk::readline_loop

pub(crate) fn readline_loop<F>(
    rl: &mut Editor<()>, 
    prompt: &str, 
    f: F
) -> Result<(), Box<dyn Error>> where
    F: FnMut(&mut Editor<()>, &str), 

Reads input lines from the user. Lines start with the string given by prompt. Each line the user enters is passed to the function f for processing.

The loop terminates (and the program ends) when EOF is reached or if an error occurs while reading the next line.