Struct chalk::program::Program

pub struct Program {
    pub(crate) type_ids: BTreeMap<Identifier, TypeKindId>,
    pub(crate) type_kinds: BTreeMap<TypeKindId, TypeKind>,
    pub(crate) struct_data: BTreeMap<StructId, Arc<StructDatum>>,
    pub(crate) impl_data: BTreeMap<ImplId, Arc<ImplDatum>>,
    pub(crate) trait_data: BTreeMap<TraitId, Arc<TraitDatum>>,
    pub(crate) associated_ty_data: BTreeMap<TypeId, Arc<AssociatedTyDatum>>,
    pub(crate) custom_clauses: Vec<ProgramClause>,
}

Fields
type_ids: BTreeMap<Identifier, TypeKindId>

From type-name to item-id. Used during lowering only.

type_kinds: BTreeMap<TypeKindId, TypeKind>

For each struct/trait:

struct_data: BTreeMap<StructId, Arc<StructDatum>>

For each struct:

impl_data: BTreeMap<ImplId, Arc<ImplDatum>>

For each impl:

trait_data: BTreeMap<TraitId, Arc<TraitDatum>>

For each trait:

associated_ty_data: BTreeMap<TypeId, Arc<AssociatedTyDatum>>

For each associated ty:

custom_clauses: Vec<ProgramClause>

For each user-specified clause

Methods

impl Program

pub(crate) fn split_projection<'p>(
&self,
projection: &'p ProjectionTy
) -> (Arc<AssociatedTyDatum>, &'p [Parameter], &'p [Parameter])

Given a projection of an associated type, split the type parameters into those that come from the trait and those that come from the associated type itself. So e.g. if you have (Iterator::Item)<F>, this would return ([F], []), since Iterator::Item is not generic and hence doesn't have any type parameters itself.

Used primarily for debugging output.

pub(crate) fn local_impl_ids(&self) -> Vec<ImplId>

Returns the ids for all impls declared in this crate.