Struct chalk::rust_ir::Program

pub struct Program {
    pub(crate) type_ids: BTreeMap<Identifier, ItemId>,
    pub(crate) type_kinds: BTreeMap<ItemId, TypeKind>,
    pub(crate) struct_data: BTreeMap<ItemId, StructDatum>,
    pub(crate) impl_data: BTreeMap<ItemId, ImplDatum>,
    pub(crate) trait_data: BTreeMap<ItemId, TraitDatum>,
    pub(crate) associated_ty_data: BTreeMap<ItemId, AssociatedTyDatum>,
    pub(crate) default_impl_data: Vec<DefaultImplDatum>,
    pub(crate) custom_clauses: Vec<ProgramClause>,
    pub(crate) lang_items: BTreeMap<LangItem, ItemId>,
}

Fields
type_ids: BTreeMap<Identifier, ItemId>

From type-name to item-id. Used during lowering only.

type_kinds: BTreeMap<ItemId, TypeKind>

For each struct/trait:

struct_data: BTreeMap<ItemId, StructDatum>

For each struct:

impl_data: BTreeMap<ItemId, ImplDatum>

For each impl:

trait_data: BTreeMap<ItemId, TraitDatum>

For each trait:

associated_ty_data: BTreeMap<ItemId, AssociatedTyDatum>

For each associated ty:

default_impl_data: Vec<DefaultImplDatum>

For each default impl (automatically generated for auto traits):

custom_clauses: Vec<ProgramClause>

For each user-specified clause

lang_items: BTreeMap<LangItem, ItemId>

Special types and traits.

Methods

impl Program

pub(crate) fn split_projection<'p>(
&self,
projection: &'p ProjectionTy
) -> (&AssociatedTyDatum, &'p [Parameter], &'p [Parameter])

Given a projection of an associated type, split the type parameters into those that come from the trait and those that come from the associated type itself. So e.g. if you have (Iterator::Item)<F>, this would return ([F], []), since Iterator::Item is not generic and hence doesn't have any type parameters itself.

Used primarily for debugging output.

impl Program

pub(super) fn visit_specializations<F>(
&self,
env: Arc<ProgramEnvironment>,
solver_choice: SolverChoice,
record_specialization: F
) -> Fallible<()> where
F: FnMut(ItemId, ItemId),

impl Program

pub(crate) fn record_specialization_priorities(
&mut self,
env: Arc<ProgramEnvironment>,
solver_choice: SolverChoice
) -> Fallible<()>

fn build_specialization_forest(
&self,
env: Arc<ProgramEnvironment>,
solver_choice: SolverChoice
) -> Fallible<Graph<ItemId, ()>>

fn set_priorities(
&mut self,
idx: NodeIndex,
forest: &Graph<ItemId, ()>,
p: usize
)

impl Program

pub fn add_default_impls(&mut self)

fn impl_provided_for(
&self,
trait_ref: TraitRef,
struct_datum: &StructDatum
) -> bool

impl Program

pub fn environment(&self) -> ProgramEnvironment